Moved hyper2json into the parser.
